Add nexus1 module with TaskPoolEntry class
The nexus1 module is partially implemented to receive
tasks, put them in in_buffer then move them to task pool
After that the load threads takes tasks from the pool and calculates
their deps count which is not implemented yet.
I added a DEBUG input to the nexus module to enable printing
the contents of buffers and internal tables whenever I send it
the right signal.